feat: Merge run and command command
