+++ lisp/help/run/command
