+++ lisp/help/run/script
